+++ lib/generators/ultimate_turbo_modal/templates/ultimate_turbo_modal.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+UltimateTurboModal.configure do |config|
+  config.flavor = FLAVOR
+  # config.close_button = true
+  # config.advance = true
+  # config.padding = true
+  # config.header = true
+  # config.header_divider = true
+  # config.footer_divider = true
+  # config.allowed_click_outside_selector = []
+end