+++ lib/ultimate_turbo_modal.rb
@@ -5,9 +5,7 @@ require "phlex/deferred_render_with_main_content"
 require "ultimate_turbo_modal/configuration"
 require "ultimate_turbo_modal/railtie"
 require "ultimate_turbo_modal/base"
-Dir[File.join(__dir__, "ultimate_turbo_modal/flavors", "*.rb")].sort.each do |file|
-  require file
-end
+require "generators/ultimate_turbo_modal/install_generator"
 
 module UltimateTurboModal
   extend self
@@ -18,6 +16,8 @@ module UltimateTurboModal
 
   def modal_class
     "UltimateTurboModal::Flavors::#{flavor.to_s.classify}".constantize
+  rescue NameError
+    raise Error, "Flavor `#{flavor.downcase}` not found. Please check your initializer file at `config/initializers/ultimate_turbo_modal.rb` and make sure to run `rails generate ultimate_turbo_modal:install`."
   end
 
   class Error < StandardError; end